beleggings_ai_agent_ultra_news/requirements.txt
pandas>=2.0
numpy>=1.24
yfinance>=0.2.52
scikit-learn>=1.3
ta>=0.11.0
PyYAML>=6.0
python-dotenv>=1.0
streamlit>=1.34
requests>=2.31
matplotlib>=3.8
pytz>=2024.1
